requirements.txt
# Core web framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2

# Machine Learning and AI
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
transformers==4.33.2
torch==2.0.1
tensorflow==2.13.0

# NLP Processing
spacy==3.6.1
nltk==3.8.1
textblob==0.17.1

# Data visualization and analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# API and data fetching
requests==2.31.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.19

# Security and encryption
cryptography==41.0.3
bcrypt==4.0.1

# Federated learning
PySyft==0.8.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0

# Production deployment
gunicorn==21.2.0
redis==4.6.0
celery==5.3.1

# Environment and configuration
python-decouple==3.8